refactor: 키 관리 역할 분리
